Split the settings' sections into their own components
